Introduce NOTPAID constant
It would be nice if there is a actual "NOTPAID" order status in the IPN Class. The reason for this is, that I´m using the IPN class payment states in a different ODM Document to track shopping cart dropouts. When a user decides to start a payment I generate a order document with a global order status. The initial state is therefore "IPN::NOTPAID". When the user drops out I have his data and generate a followup. If the payment later gets confirmed I change this to IPN::PAID and so on ...

Also I think filling the orderStatus with an initial state is better than testing for orderStatus != null