We only add \r\n when Paypal sends us a vanilla \n
Paypal is sending \r\n on our website. With the current code, \n are replaced by \r\n all the time, resulting in \r\r\n. This gives us an INVALID IPN response, even when it was valid. By adding a check to make sure Paypal is sending vanilla \n before replacing it with a \r\n, we solved this bug once and for all.